==> CellCycleHela/02_tracesProcessing_improved.R <==
source("/Users/isabell/Desktop/projects/ProteoformProject/CCprofilerAnalysis/thesis/CellCycle_header.R")

#' ## Load data
pepTracesRaw <- readRDS("pepTracesRaw.rda")
design_matrix <- readRDS("design_matrix.rda")

#' ## Reduce traces to a single representative peptide for groups of peptides generated by multiple missed cleavages
#'
#' ##### Integrate traces across conditions to perform the representative peptide selection consistently across conditions
traces_integrated <- integrateTraceIntensities(pepTracesRaw,
                                               design_matrix = NULL,
                                               integrate_within = NULL,
                                               aggr_fun = "sum")

#' ## Annotate peptide sequence positions
traces_integrated_iso_pos <- annotatePeptideSequences(traces_integrated, fasta_file="../InputData/human.fasta")
saveRDS(traces_integrated_iso_pos, "traces_integrated_iso_pos.rds")

#' ## Select representative peptides
traces_integrated_start_end <- summarizeAlternativePeptideSequences(
  traces_integrated_iso_pos, topN=1,start="PeptidePositionStart",end="PeptidePositionEnd", verbose=F)

saveRDS(traces_integrated_start_end, "traces_integrated_start_end.rds")

#' ## Subset traces list to selected peptides
pepTracesRepresentativePeps <- subset(pepTracesRaw,
                                      unique(traces_integrated_start_end$traces$id))

common_names <- names(pepTracesRepresentativePeps$Interphase1$trace_annotation)[
  which(names(pepTracesRepresentativePeps$Interphase1$trace_annotation) 
        %in% names(traces_integrated_start_end$trace_annotation))]

for(n in names(pepTracesRepresentativePeps)){
  pepTracesRepresentativePeps[[n]]$trace_annotation <- merge(pepTracesRepresentativePeps[[n]]$trace_annotation, 
                                                             traces_integrated_start_end$trace_annotation, by=common_names, all.x=T, all.y=F, sort=F)
}

valid_pep_ids = traces_integrated_start_end$trace_annotation[which(! is.na(PeptidePositionStart))]$id
pepTracesRepresentativePeps <- subset(pepTracesRepresentativePeps, valid_pep_ids)
saveRDS(pepTracesRepresentativePeps, "pepTracesRepresentativePeps.rds")


#' ## Missing value detection and imputation
#' Convert zeros in missing value locations to NA:
pepTracesMV <- findMissingValues(pepTracesRepresentativePeps,
                                 bound_left = 1,
                                 bound_right = 1,
                                 consider_borders = TRUE)

#' Impute NA values by fitting a spline:
pepTracesImp <- imputeMissingVals(pepTracesMV, method = "spline")

#' Plot imputation summary:
plotImputationSummary(pepTracesMV, pepTracesImp, PDF = T, plot_traces = T, max_n_traces = 2)

###################################
#' IntegrateAllTraces for filtering
###################################
traces_combined <- combineTracesMutiCond(pepTracesImp)

#' ## Filter by consecutive IDs
pepTracesConsIds <- filterConsecutiveIdStretches(traces_combined,
                                             min_stretch_length = 3,
                                             remove_empty = T)

saveRDS(pepTracesConsIds, "pepTracesConsIds.rds")

#' ## Remove traces with 0 standard deviation (can't be clustered)
zerovar <- apply(getIntensityMatrix(pepTracesConsIds), 1, var) > 0
pepTracesZerovar <- subset.traces(pepTracesConsIds,
                                trace_subset_ids = names(zerovar[zerovar]))

#' ## Remove single peptide genes
pepTracesConsIds_multiPep <- filterSinglePeptideHits(pepTracesZerovar)

saveRDS(pepTracesConsIds_multiPep, "pepTracesConsIds_multiPep.rds")

#' ## Remove outlier peptides
#' Outlier peptides are defined as peptides that do not have any high correlating
#' sibling peptide. Genes with only a single peptide are automatically removed.
# pepTraces_maxCorr <- filterByMaxCorr(pepTracesConsIds_multiPep,
#                                  cutoff = 0.7, plot = T, PDF=T, name="maxCorrHist")
## We removed this outlier filtering, after the implementation of an automatic outlier 
# removal during the clustering step. 
pepTraces_maxCorr <- pepTracesConsIds_multiPep

validPeps <- unique(pepTraces_maxCorr$trace_annotation$id)
pepTracesList_filtered <- lapply(pepTracesImp, function(x){
  subset(x, trace_subset_ids=validPeps)
})
class(pepTracesList_filtered) <- "tracesList"

#' Update fraction annotation:
pepTracesList_filtered <- updateTraces(pepTracesList_filtered)

#' ## Save objects
saveRDS(pepTracesList_filtered, "pepTracesList_filtered.rda")

#' ## QC plots
test_proteins <- c("P06493")
pepTest <- subset(pepTracesList_filtered, trace_subset_ids = test_proteins, trace_subset_type = "protein_id")
plot(pepTest, design_matrix = design_matrix, log = FALSE, legend = T, PDF = TRUE,
     name = paste0("FILTERED_PeptideTraces_CDK1_","P06493"), isoformAnnotation = F, plot = TRUE, highlight = NULL, highlight_col = NULL, colour_by = "Entry_name")
